PostFiltering.h created.
Many times we need to use this from somewhere else.
